Changelog (2024-06-18 : 00) DONE - Changed default settings for CommonUI, was set to use default input config which was messing things up with user focus, was causing issues after having activated a commonui activatable widget. - Fix missing or incorrect tablerow type, needed to add some context strings to the GetRow calls to see what I have overlooked - Fixed a selection bug related to the main menu - Added a funciotn in the builder subsystem name 'GetBuildableWithActionsFromClassStatic' which as the name implies returns a FPDBuildable based on teh given class, but only if said buildable has any action-contexts associated with it. This is made use of to avoid triggering the actions menu to open when dbl clicking a buildable with no actions. - Noticed and resolved some widget animations that were being triggered by incorrect functions, and added some more safety related to calling animatiosn with 0 second of duration - Had forgotten to assign the DestroyBuildable action to the contexts applied to WorkerHut0, was causing it to not display the 'destroy buildable' actions
IN PROGRESS

- (BUGS -- RESOLVE BY TOMORROW!) AS I was wrapping up I found 3 bugs related to many recent changes in selection as complexity increased to handle different types of selections:
BUG 1: When opening a entity workers build menu there is no way to close it anymore, it does not get back to normal input mode,
	double clicking anywhere does not remove it anymore.
BUG 2: If removing world actors fast eough it fails to actually remove them. Might need to set up a World actor deletion queue in the widget so if we spam deletions it never misses to delete one (Didn't I already set this up?? Will need to check the subsystems in the RTS plugin might already have it in place and am forgetting to call it)

BUG 3: Selecting a entity to drag to location causes the appropriate path, but now the path doesn't stay on target when having selected a target. Entity walks towards target as expected but the visual path indicator just sticks to the cursor, as if a target was never selected

TODO
- Handling unit spawn actions: In 'AGodHandPawn::SelectActionMenuEntry_Implementation':
	// @todo 1. Need to select a entity config ased on the spawn tag and use it to spawn an entity via the spawn manager
	// @todo 2. Will need some mapping between unit-type tags and their entity configs, multiple unit-type tags may share the same configs
	// 	2.a: ^ (Put this in FPDBuildWorker, a new property for entity configs)
	// @todo 3. Spawn here now that have validated that the worker is valid type
	// @todo 4. need a spawn queue and unit spawn rules, timings and such

- Add some icons for the differn buildable context, buildable entries as-well as for the buildable action contexts and buildable actions therein